[Refactor] 지도 핀 마커 컴포넌트 리팩터링 #32
