+++ src/components/pages/map/index.ts
@@ -3,3 +3,4 @@ export { default as Error } from './Error';
 export { default as Category } from './Category';
 export { default as CurrentLocationButton } from './CurrentLocationButton';
 export { default as Header } from './Header';
+export { default as Marker } from './MapMarker';
