+++ src/hooks/pages/map/index.ts
@@ -1,0 +1,2 @@
+export { default as useMap } from './useMap';
+export { default as useBottomMenu } from './useBottomMenu';
